+++ zcash_client_backend/src/data_api.rs
@@ -2,6 +2,7 @@
 
 use std::collections::HashMap;
 use std::fmt::Debug;
+use std::num::NonZeroU32;
 use std::{cmp, ops::Range};
 
 use incrementalmerkletree::Retention;
@@ -97,7 +98,7 @@ pub trait WalletRead {
     /// This will return `Ok(None)` if no block data is present in the database.
     fn get_target_and_anchor_heights(
         &self,
-        min_confirmations: u32,
+        min_confirmations: NonZeroU32,
     ) -> Result<Option<(BlockHeight, BlockHeight)>, Self::Error> {
         self.block_height_extrema().map(|heights| {
             heights.map(|(min_height, max_height)| {
@@ -106,7 +107,7 @@ pub trait WalletRead {
                 // Select an anchor min_confirmations back from the target block,
                 // unless that would be before the earliest block we have.
                 let anchor_height = BlockHeight::from(cmp::max(
-                    u32::from(target_height).saturating_sub(min_confirmations),
+                    u32::from(target_height).saturating_sub(min_confirmations.into()),
                     u32::from(min_height),
                 ));